% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simplification.R
\name{get.data}
\alias{get.data}
\title{Get the list of data points.}
\usage{
get.data(x)
}
\arguments{
\item{x}{An object of class simplification, created through
\code{initialize.simplification}}
}
\value{
A #data-by-4 matrix, where the first three columns
			stores the coordinates of the data points, while
			the data values are reported in the fourth column.
}
\description{
Get the list of data points.
}
