% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simplification.R
\name{get.elems}
\alias{get.elems}
\title{Get the list of elements.}
\usage{
get.elems(x)
}
\arguments{
\item{x}{An object of class simplification, created through
\code{initialize.simplification}}
}
\value{
A #elements-by-3 matrix, where the i-th row	stores
			the Id's of the vertices of the i-th triangle.
}
\description{
Get the list of elements.
}
