% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simplification.R
\name{get.nodes}
\alias{get.nodes}
\title{Get the list of nodes.}
\usage{
get.nodes(x)
}
\arguments{
\item{x}{An object of class simplification, created through
\code{initialize.simplification}}
}
\value{
A #nodes-by-3 matrix, where the i-th row stores
			the coordinates of the i-th node.
}
\description{
Get the list of nodes.
}
