% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simplification.R
\name{get.quantity.of.information}
\alias{get.quantity.of.information}
\title{Get the quantity of information for each element.}
\usage{
get.quantity.of.information(x)
}
\arguments{
\item{x}{An object of class simplification, created through
\code{initialize.simplification}}
}
\value{
A #elements-by-1 vector, storing the quantity of
			information for each element.
}
\description{
Get the quantity of information for each element.
}
