% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simplification.R
\name{initialize.simplification}
\alias{initialize.simplification}
\title{Create a simplification structure}
\usage{
initialize.simplification(file, wgeom = 1/3, wdisp = 1/3, wequi = 1/3)
}
\arguments{
\item{file}{A string reporting the absolute or relative path to the input mesh;
.inp and .vtk file formats are supported.}

\item{wgeom}{Weight for the geometric component of the edge cost function;
default is 1/3. Note that the all weights should be positive and
    sum up to one.}

\item{wdisp}{Weight for the data displacement component of the edge cost function;
default is 1/3. Note that the all weights should be positive and
    sum up to one.}

\item{wequi}{Weight for the data equidistribution component of the edge cost function;
default is 1/3. Note that the all weights should be positive and
    sum up to one.}
}
\value{
An object of class simplification, provided with the following fields:
				\item{\code{simplifier}} An object of class \code{RcppSimplification}
}
\description{
This function instantiates a class storing all informations regarding
				a surface two dimensional grid and useful to run the simplification process.
				This function heavily relies on \code{RcppSimplification} - a wrapper for the
				template class \code{simplification<Triangle, MeshType::DATA, DataGeo>} 
				provided by the C++ library \code{meshsimplification}.
}
