% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simplification.R
\name{run.simplification}
\alias{run.simplification}
\title{Run the simplification process.}
\usage{
run.simplification(x, numNodesMax, file = '')
}
\arguments{
\item{x}{An object of class \code{simplification}, created through 
\code{initialize.simplification}}

\item{numNodesMax}{Final number of nodes.}

\item{file}{String specifying the path to the location where the decimated
mesh will be stored; .inp file format supported. If the path
is not provided, the mesh will no printed.}
}
\description{
Function running the simplification process by invoking the 
					method \code{simplificate} of the \code{RcppSimplification}
					through the attribute \code{simplifier} of the \code{simplification}
					object given as argument. Note that the procedure is completely 
					carried out at the C++ level by the \code{meshsimplification} 
					library.
}
\examples{
## Instantiate a simplification object for the simplification of pawn geometry
## Suppose that the mesh file is placed in the R working folder
## Components of the edge cost function are equally weighted
obj <- initialize.simplification('pawn.inp')
## Run the simplification strategy, reducing the mesh down to n = 1000 nodes
## Once the procedure is over, print the mesh to file
run.simplification(1000, 'pawn_simplified.inp')
}
